requirements.txt
Flask==2.3.3
reportlab==4.0.4
PyPDF2==3.0.1
Werkzeug==2.3.7
python-dotenv==1.0.0
PyMuPDF==1.23.5
fpdf2==2.7.8
